refactor: rename to clientTelemetry, remove connections from telemetry, move repo impl to telemetry.go
