+++ journal/middleware/tracing.go
@@ -46,7 +46,7 @@ func (tm *tracing) RetrieveAll(ctx context.Context, session smqauthn.Session, pa
 	return tm.svc.RetrieveAll(ctx, session, page)
 }
 
-func (tm *tracing) RetrieveClientTelemetry(ctx context.Context, session smqauthn.Session, clientID string) (j journal.ClientsTelemetry, err error) {
+func (tm *tracing) RetrieveClientTelemetry(ctx context.Context, session smqauthn.Session, clientID string) (j journal.ClientTelemetry, err error) {
 	ctx, span := tm.tracer.Start(ctx, "retrieve", trace.WithAttributes(
 		attribute.String("client_id", clientID),
 		attribute.String("domain_id", session.DomainID),
