+++ journal/postgres/journal.go
@@ -5,7 +5,6 @@ package postgres
 
 import (
 	"context"
-	"database/sql"
 	"encoding/json"
 	"fmt"
 	"strings"
@@ -15,7 +14,6 @@ import (
 	"github.com/absmach/supermq/pkg/errors"
 	repoerr "github.com/absmach/supermq/pkg/errors/repository"
 	"github.com/absmach/supermq/pkg/postgres"
-	"github.com/jackc/pgtype"
 )
 
 type repository struct {
@@ -104,71 +102,6 @@ func (repo *repository) RetrieveAll(ctx context.Context, page journal.Page) (jou
 	return journalsPage, nil
 }
 
-func (repo *repository) SaveClientTelemetry(ctx context.Context, ct journal.ClientsTelemetry) error {
-	q := `INSERT INTO clients_telemetry (client_id, domain_id, connections, messages, subscriptions, first_seen, last_seen)
-		VALUES (:client_id, :domain_id, :connections, :messages, :subscriptions, :first_seen, :last_seen);`
-
-	dbct, err := toDBClientsTelemetry(ct)
-	if err != nil {
-		return errors.Wrap(repoerr.ErrCreateEntity, err)
-	}
-
-	if _, err := repo.db.NamedExecContext(ctx, q, dbct); err != nil {
-		return postgres.HandleError(repoerr.ErrCreateEntity, err)
-	}
-
-	return nil
-}
-
-func (repo *repository) DeleteClientTelemetry(ctx context.Context, clientID, domainID string) error {
-	q := "DELETE FROM clients_telemetry AS ct WHERE ct.client_id = :client_id AND ct.domain_id = :domain_id;"
-
-	dbct := dbClientsTelemetry{
-		ClientID: clientID,
-		DomainID: domainID,
-	}
-
-	result, err := repo.db.NamedExecContext(ctx, q, dbct)
-	if err != nil {
-		return postgres.HandleError(repoerr.ErrRemoveEntity, err)
-	}
-	if rows, _ := result.RowsAffected(); rows == 0 {
-		return repoerr.ErrNotFound
-	}
-	return nil
-}
-
-func (repo *repository) RetrieveClientTelemetry(ctx context.Context, clientID, domainID string) (journal.ClientsTelemetry, error) {
-	q := "SELECT * FROM clients_telemetry WHERE client_id = :client_id AND domain_id = :domain_id;"
-
-	dbct := dbClientsTelemetry{
-		ClientID: clientID,
-		DomainID: domainID,
-	}
-
-	rows, err := repo.db.NamedQueryContext(ctx, q, dbct)
-	if err != nil {
-		return journal.ClientsTelemetry{}, postgres.HandleError(repoerr.ErrViewEntity, err)
-	}
-	defer rows.Close()
-
-	dbct = dbClientsTelemetry{}
-	if rows.Next() {
-		if err = rows.StructScan(&dbct); err != nil {
-			return journal.ClientsTelemetry{}, postgres.HandleError(repoerr.ErrViewEntity, err)
-		}
-
-		ct, err := toClientsTelemetry(dbct)
-		if err != nil {
-			return journal.ClientsTelemetry{}, errors.Wrap(repoerr.ErrFailedOpDB, err)
-		}
-
-		return ct, nil
-	}
-
-	return journal.ClientsTelemetry{}, repoerr.ErrNotFound
-}
-
 func pageQuery(pm journal.Page) string {
 	var query []string
 	var emq string
@@ -255,72 +188,5 @@ func toJournal(dbj dbJournal) (journal.Journal, error) {
 		OccurredAt: dbj.OccurredAt,
 		Attributes: attributes,
 		Metadata:   metadata,
-	}, nil
-}
-
-type dbClientsTelemetry struct {
-	ClientID         string           `db:"client_id"`
-	DomainID         string           `db:"domain_id"`
-	Connections      pgtype.TextArray `db:"connections"`
-	Subscriptions    pgtype.TextArray `db:"subscriptions"`
-	InboundMessages  uint64           `db:"inbound_messages"`
-	OutboundMessages uint64           `db:"outbound_messages"`
-	FirstSeen        time.Time        `db:"first_seen"`
-	LastSeen         sql.NullTime     `db:"last_seen"`
-}
-
-func toDBClientsTelemetry(ct journal.ClientsTelemetry) (dbClientsTelemetry, error) {
-	var conns pgtype.TextArray
-	if err := conns.Set(ct.Connections); err != nil {
-		return dbClientsTelemetry{}, err
-	}
-
-	var subs pgtype.TextArray
-	if err := subs.Set(ct.Subscriptions); err != nil {
-		return dbClientsTelemetry{}, err
-	}
-
-	var lastSeen sql.NullTime
-	if ct.LastSeen != (time.Time{}) {
-		lastSeen = sql.NullTime{Time: ct.LastSeen, Valid: true}
-	}
-
-	return dbClientsTelemetry{
-		ClientID:         ct.ClientID,
-		DomainID:         ct.DomainID,
-		Connections:      conns,
-		Subscriptions:    subs,
-		InboundMessages:  ct.InboundMessages,
-		OutboundMessages: ct.OutboundMessages,
-		FirstSeen:        ct.FirstSeen,
-		LastSeen:         lastSeen,
-	}, nil
-}
-
-func toClientsTelemetry(dbct dbClientsTelemetry) (journal.ClientsTelemetry, error) {
-	var conns []string
-	for _, e := range dbct.Connections.Elements {
-		conns = append(conns, e.String)
-	}
-
-	var subs []string
-	for _, e := range dbct.Subscriptions.Elements {
-		subs = append(subs, e.String)
-	}
-
-	var lastSeen time.Time
-	if dbct.LastSeen.Valid {
-		lastSeen = dbct.LastSeen.Time
-	}
-
-	return journal.ClientsTelemetry{
-		ClientID:         dbct.ClientID,
-		DomainID:         dbct.DomainID,
-		Connections:      conns,
-		Subscriptions:    subs,
-		InboundMessages:  dbct.InboundMessages,
-		OutboundMessages: dbct.OutboundMessages,
-		FirstSeen:        dbct.FirstSeen,
-		LastSeen:         lastSeen,
 	}, nil
 }
